requirements.txt
altgraph==0.17.4
blinker==1.8.1
bottle==0.12.25
cffi==1.16.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.3
importlib_metadata==7.1.0
isoweek==1.3.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.0
pefile==2023.2.7
pip==23.2.1
proxy_tools==0.1.0
pyarrow==15.0.0
pycparser==2.21
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pythonnet==3.0.3
pytz==2024.1
pywebview==4.4.1
pywin32==306
pywin32-ctypes==0.2.2
pyxlsb==1.0.10
setuptools==70.0.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.2
XlsxWriter==3.1.9
zipp==3.18.1
